Split RawDisplay code into 3 files
The splitting prepares the code for inclusion into O2. The functionality is
essentially the same as before.
